+++ wingle/src/main/java/kr/co/wingle/hello/HelloRequest.java
@@ -1,0 +1,11 @@
+package kr.co.wingle.hello;
+
+import javax.validation.constraints.NotBlank;
+
+import lombok.Getter;
+
+@Getter
+public class HelloRequest {
+	@NotBlank(message = "이름이 없습니다.")
+	private String name;
+}
